video: tegra: hdmi: fix audio to work with any pixel clock rate
The audio setting implementation was limited to a few specific pixel
clocks. This prevented HDMI audio from working on several test devices
as they need a pixel clock that is not supported by this implementation.

Fix this by implementing the algorithm provided in the TRM using fixed
point arithmetic. This allows the driver to cope with any sane pixel
clock rate and removes the need to round the pixel clock to one on the
rate that allow supporting audio.

Change-Id: I569a2556e43bcc6f6078c8759a14a33d7b18ff27
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2647
Reviewed-by: Julian Scheel <julian@jusst.de>